refactor: type filenames to Types
