+++ app/(tabs)/addRide.tsx
@@ -22,7 +22,7 @@ import useVehicleTypes from '@/hooks/data/useVehicleTypes'
 import { useToggleLoading } from '@/hooks/useLoading'
 import useModalHandler from '@/hooks/useModalHandler'
 import { inputElementStyles } from '@/src/styles/InputStyles'
-import { AddableLap, AddableTravel } from '@/src/types/AddableTravels'
+import { AddableLap, AddableTravel } from '@/src/types/AddableTypes'
 import { datetimeFieldToCapitals, formatDateForDisplay } from '@/src/utils/utils'
 import { router, useFocusEffect } from 'expo-router'
 import moment from 'moment-timezone'
