+++ app/(tabs)/main/editRide.tsx
@@ -24,8 +24,8 @@ import useVehicleTypes from '@/hooks/data/useVehicleTypes'
 import { useLoading } from '@/hooks/useLoading'
 import useModalHandler from '@/hooks/useModalHandler'
 import { inputElementStyles } from '@/src/styles/InputStyles'
-import { AddableLap } from '@/src/types/AddableTravels'
-import { EditableTravel } from '@/src/types/EditableTravels'
+import { AddableLap } from '@/src/types/AddableTypes'
+import { EditableTravel } from '@/src/types/EditableTypes'
 import { getDateToIsoString } from '@/src/utils/dateUtils'
 import { datetimeFieldToCapitals, formatDateForDisplay } from '@/src/utils/utils'
 import { router, useFocusEffect } from 'expo-router'
