+++ app/(tabs)/main/travelDetail.tsx
@@ -13,8 +13,8 @@ import useVehicleTypes from '@/hooks/data/useVehicleTypes'
 import useTravelDetail from '@/hooks/useTravelDetail'
 import { colors } from '@/src/const/color'
 import { travelDetailStyles } from '@/src/styles/TravelDetailStyles'
-import { CompleteRide } from '@/src/types/CompleteTravels'
-import { Stop } from '@/src/types/Travels'
+import { CompleteRide } from '@/src/types/CompleteTypes'
+import { Stop } from '@/src/types/Types'
 import { formatMsToMinutes, sumTimesToMs } from '@/src/utils/dateUtils'
 import { getSimpleCentroid } from '@/src/utils/mapUtils'
 import { MarkerView } from '@maplibre/maplibre-react-native'
