+++ components/input/VehicleSelector.tsx
@@ -1,7 +1,7 @@
 import { useTheme } from "@/context/ThemeContext"
 import { colors } from "@/src/const/color"
-import { CompleteVehicleType } from "@/src/types/CompleteTravels"
-import { IconType } from "@/src/types/Travels"
+import { CompleteVehicleType } from "@/src/types/CompleteTypes"
+import { IconType } from "@/src/types/Types"
 import { TouchableOpacity } from "react-native"
 import CustomIcon from "../CustomIcon"
 import Input from "./Input"
