+++ components/modal/addModal/AddDirectionModal.tsx
@@ -2,7 +2,7 @@ import Button from "@/components/button/BaseButton"
 import Input from "@/components/input/Input"
 import { TextInputBlock } from "@/components/input/TextInput"
 import { useDialog } from "@/context/DialogContext"
-import { AddableDirection } from "@/src/types/AddableTravels"
+import { AddableDirection } from "@/src/types/AddableTypes"
 import { BaseModalContentProps } from "@/src/types/ModalContentProps"
 import { useState } from "react"
 import { View } from "react-native"
