+++ components/modal/addModal/AddIconModal.tsx
@@ -5,7 +5,7 @@ import { TextInputBase } from "@/components/input/TextInput"
 import { useDialog } from "@/context/DialogContext"
 import { useTheme } from "@/context/ThemeContext"
 import { inputElementStyles } from "@/src/styles/InputStyles"
-import { AddableIconType } from "@/src/types/AddableTravels"
+import { AddableIconType } from "@/src/types/AddableTypes"
 import { BaseModalContentProps } from "@/src/types/ModalContentProps"
 import { useState } from "react"
 import { View } from "react-native"
