+++ components/modal/addModal/AddLapModal.tsx
@@ -11,7 +11,7 @@ import { useTheme } from '@/context/ThemeContext'
 import useLocation from '@/hooks/useLocation'
 import useModalHandler from '@/hooks/useModalHandler'
 import { inputElementStyles } from '@/src/styles/InputStyles'
-import { AddableLap, AddableLapModalProp } from '@/src/types/AddableTravels'
+import { AddableLap, AddableLapModalProp } from '@/src/types/AddableTypes'
 import { getDateToIsoString } from '@/src/utils/dateUtils'
 import { formatDateForDisplay, formatLapTimeDisplay } from '@/src/utils/utils'
 import { UserLocation } from '@maplibre/maplibre-react-native'
