+++ components/modal/addModal/AddRouteModal.tsx
@@ -9,8 +9,8 @@ import { useTheme } from "@/context/ThemeContext"
 import useVehicleTypes from "@/hooks/data/useVehicleTypes"
 import useModalHandler from "@/hooks/useModalHandler"
 import { inputElementStyles } from "@/src/styles/InputStyles"
-import { AddableRoute } from "@/src/types/AddableTravels"
-import { ModalProp } from "@/src/types/TravelModal"
+import { AddableRoute } from "@/src/types/AddableTypes"
+import { ModalProp } from "@/src/types/ModalTypes"
 import { useFocusEffect } from "expo-router"
 import { useCallback, useState } from "react"
 import { ScrollView, View } from "react-native"
