+++ components/modal/addModal/AddStopModal.tsx
@@ -8,7 +8,7 @@ import { useTheme } from "@/context/ThemeContext"
 import useVehicleTypes from "@/hooks/data/useVehicleTypes"
 import useModalHandler from "@/hooks/useModalHandler"
 import { inputElementStyles } from "@/src/styles/InputStyles"
-import { AddableCoordinates, AddableStop } from "@/src/types/AddableTravels"
+import { AddableCoordinates, AddableStop } from "@/src/types/AddableTypes"
 import { BaseModalContentProps } from "@/src/types/ModalContentProps"
 import { useState } from "react"
 import { ScrollView, View } from "react-native"
