+++ components/modal/addModal/AddVehicleTypeModal.tsx
@@ -5,8 +5,8 @@ import { IconSelector } from "@/components/input/VehicleSelector"
 import { useDialog } from "@/context/DialogContext"
 import { useTheme } from "@/context/ThemeContext"
 import { inputElementStyles } from "@/src/styles/InputStyles"
-import { AddableVehicleType } from "@/src/types/AddableTravels"
-import { VehicleTypeModalProp } from "@/src/types/TravelModal"
+import { AddableVehicleType } from "@/src/types/AddableTypes"
+import { VehicleTypeModalProp } from "@/src/types/ModalTypes"
 import { useState } from "react"
 import { ScrollView, View } from "react-native"
