+++ components/modal/editModal/EditIconModal.tsx
@@ -6,7 +6,7 @@ import { useDataEditContext } from "@/context/DataEditContext"
 import { useDialog } from "@/context/DialogContext"
 import { useTheme } from "@/context/ThemeContext"
 import { inputElementStyles } from "@/src/styles/InputStyles"
-import { AddableIconType } from "@/src/types/AddableTravels"
+import { AddableIconType } from "@/src/types/AddableTypes"
 import { BaseModalContentProps } from "@/src/types/ModalContentProps"
 import { useState } from "react"
 import { View } from "react-native"
