+++ components/modal/editModal/EditLapModal.tsx
@@ -10,7 +10,7 @@ import { useDialog } from '@/context/DialogContext'
 import { useTheme } from '@/context/ThemeContext'
 import useModalHandler from '@/hooks/useModalHandler'
 import { inputElementStyles } from '@/src/styles/InputStyles'
-import { EditableLapModalProp } from '@/src/types/EditableTravels'
+import { EditableLapModalProp } from '@/src/types/EditableTypes'
 import { getDateToIsoString } from '@/src/utils/dateUtils'
 import { formatDateForDisplay } from '@/src/utils/utils'
 import React, { useEffect, useState } from 'react'
