+++ components/modal/editModal/EditRouteModal.tsx
@@ -10,9 +10,9 @@ import { useTheme } from "@/context/ThemeContext"
 import useVehicleTypes from "@/hooks/data/useVehicleTypes"
 import useModalHandler from "@/hooks/useModalHandler"
 import { inputElementStyles } from "@/src/styles/InputStyles"
-import { CompleteVehicleType } from "@/src/types/CompleteTravels"
-import { EditableRoute } from "@/src/types/EditableTravels"
-import { ModalProp } from "@/src/types/TravelModal"
+import { CompleteVehicleType } from "@/src/types/CompleteTypes"
+import { EditableRoute } from "@/src/types/EditableTypes"
+import { ModalProp } from "@/src/types/ModalTypes"
 import { sortByIdToFront } from "@/src/utils/utils"
 import { useFocusEffect } from "expo-router"
 import { useCallback, useEffect, useRef, useState } from "react"
