+++ components/modal/editModal/EditStopModal.tsx
@@ -10,9 +10,9 @@ import useStopsVehicleTypes from "@/hooks/data/useStopVehicleTypes"
 import useVehicleTypes from "@/hooks/data/useVehicleTypes"
 import useModalHandler from "@/hooks/useModalHandler"
 import { inputElementStyles } from "@/src/styles/InputStyles"
-import { AddableCoordinates } from "@/src/types/AddableTravels"
-import { CompleteVehicleType } from "@/src/types/CompleteTravels"
-import { EditableStop } from "@/src/types/EditableTravels"
+import { AddableCoordinates } from "@/src/types/AddableTypes"
+import { CompleteVehicleType } from "@/src/types/CompleteTypes"
+import { EditableStop } from "@/src/types/EditableTypes"
 import { BaseModalContentProps } from "@/src/types/ModalContentProps"
 import { sortByIdToFront } from "@/src/utils/utils"
 import { useEffect, useState } from "react"
