+++ components/modal/editModal/EditVehicleTypeModal.tsx
@@ -7,9 +7,9 @@ import { useDialog } from "@/context/DialogContext"
 import { useTheme } from "@/context/ThemeContext"
 import useIcons from "@/hooks/data/useIcons"
 import { inputElementStyles } from "@/src/styles/InputStyles"
-import { EditableVehicleType } from "@/src/types/EditableTravels"
+import { EditableVehicleType } from "@/src/types/EditableTypes"
 import { BaseModalContentProps } from "@/src/types/ModalContentProps"
-import { IconType } from "@/src/types/Travels"
+import { IconType } from "@/src/types/Types"
 import { sortByIdToFront } from "@/src/utils/utils"
 import { useEffect, useRef, useState } from "react"
 import { ScrollView, View } from "react-native"
