+++ components/modal/rideModal/AddRideLapsModal.tsx
@@ -4,7 +4,7 @@ import ModalTemplate from '@/components/ModalTemplate'
 import { useTheme } from '@/context/ThemeContext'
 import useModalHandler from '@/hooks/useModalHandler'
 import { modalStyles } from '@/src/styles/ModalStyles'
-import { AddableLap, AddableLapsModalProp } from '@/src/types/AddableTravels'
+import { AddableLap, AddableLapsModalProp } from '@/src/types/AddableTypes'
 import { LocationManager } from '@maplibre/maplibre-react-native'
 import { useFocusEffect } from 'expo-router'
 import React, { useCallback, useEffect, useState } from 'react'
