+++ components/modal/rideModal/EditRideDirectionModal.tsx
@@ -3,7 +3,7 @@ import { TextInputBase } from "@/components/input/TextInput"
 import ModalTemplate from "@/components/ModalTemplate"
 import { useTheme } from "@/context/ThemeContext"
 import { modalElementStyles, modalStyles } from "@/src/styles/ModalStyles"
-import { EditableTravelDirectionModalProp } from "@/src/types/EditableTravels"
+import { EditableTravelDirectionModalProp } from "@/src/types/EditableTypes"
 import { useMemo } from "react"
 import { Pressable, View } from "react-native"
 import FlatlistBase from "../FlatlistPicker"
