+++ components/modal/rideModal/EditRideLapsModal.tsx
@@ -4,7 +4,7 @@ import ModalTemplate from '@/components/ModalTemplate'
 import { useTheme } from '@/context/ThemeContext'
 import useModalHandler from '@/hooks/useModalHandler'
 import { modalStyles } from '@/src/styles/ModalStyles'
-import { EditableLapsModalProp } from '@/src/types/EditableTravels'
+import { EditableLapsModalProp } from '@/src/types/EditableTypes'
 import { sortLaps } from '@/src/utils/dataUtils'
 import { LocationManager } from '@maplibre/maplibre-react-native'
 import { useFocusEffect } from 'expo-router'
