+++ components/modal/rideModal/EditRideRouteModal.tsx
@@ -3,8 +3,8 @@ import { TextInputBase } from "@/components/input/TextInput"
 import ModalTemplate from "@/components/ModalTemplate"
 import { useTheme } from "@/context/ThemeContext"
 import { modalElementStyles, modalStyles } from "@/src/styles/ModalStyles"
-import { EditableTravelRouteModalProp } from "@/src/types/EditableTravels"
-import { Route } from "@/src/types/Travels"
+import { EditableTravelRouteModalProp } from "@/src/types/EditableTypes"
+import { Route } from "@/src/types/Types"
 import { useMemo } from "react"
 import { Pressable, View } from "react-native"
 import FlatlistBase from "../FlatlistPicker"
