+++ components/modal/rideModal/EditRideStopModal.tsx
@@ -4,8 +4,8 @@ import { TextInputBase } from "@/components/input/TextInput"
 import ModalTemplate from "@/components/ModalTemplate"
 import { useTheme } from "@/context/ThemeContext"
 import { modalElementStyles, modalStyles } from "@/src/styles/ModalStyles"
-import { CompleteStop } from "@/src/types/CompleteTravels"
-import { EditableTravelStopModalProp } from "@/src/types/EditableTravels"
+import { CompleteStop } from "@/src/types/CompleteTypes"
+import { EditableTravelStopModalProp } from "@/src/types/EditableTypes"
 import { useEffect, useMemo, useState } from "react"
 import { Pressable, View } from "react-native"
 import FlatlistBase from "../FlatlistPicker"
