+++ hooks/data/useLaps.tsx
@@ -1,8 +1,8 @@
 import { ManageableLap } from "@/components/modal/FlatlistPicker"
 import { db } from "@/src/services/dataDbService"
-import { AddableLap } from "@/src/types/AddableTravels"
-import { CompleteLap } from "@/src/types/CompleteTravels"
-import { EditableLap } from "@/src/types/EditableTravels"
+import { AddableLap } from "@/src/types/AddableTypes"
+import { CompleteLap } from "@/src/types/CompleteTypes"
+import { EditableLap } from "@/src/types/EditableTypes"
 import { groupLapsWithStop } from "@/src/utils/groupingUtils"
 import { SQLBatchTuple } from "@op-engineering/op-sqlite"
 import { useEffect, useState } from "react"
