+++ hooks/data/useRoutes.tsx
@@ -1,8 +1,8 @@
 import { db } from "@/src/services/dataDbService"
-import { AddableRoute } from "@/src/types/AddableTravels"
-import { CompleteRoute } from "@/src/types/CompleteTravels"
-import { EditableRoute } from "@/src/types/EditableTravels"
-import { Route } from "@/src/types/Travels"
+import { AddableRoute } from "@/src/types/AddableTypes"
+import { CompleteRoute } from "@/src/types/CompleteTypes"
+import { EditableRoute } from "@/src/types/EditableTypes"
+import { Route } from "@/src/types/Types"
 import { groupRoutesWithVehicleTypes } from "@/src/utils/groupingUtils"
 import { SQLBatchTuple } from "@op-engineering/op-sqlite"
 import { useEffect, useState } from "react"
