+++ hooks/data/useStops.tsx
@@ -1,8 +1,8 @@
 import { db } from "@/src/services/dataDbService"
-import { AddableStop } from "@/src/types/AddableTravels"
-import { CompleteStop } from "@/src/types/CompleteTravels"
-import { EditableStop } from "@/src/types/EditableTravels"
-import { Stop } from "@/src/types/Travels"
+import { AddableStop } from "@/src/types/AddableTypes"
+import { CompleteStop } from "@/src/types/CompleteTypes"
+import { EditableStop } from "@/src/types/EditableTypes"
+import { Stop } from "@/src/types/Types"
 import { groupStopsWithVehicleTypes } from "@/src/utils/groupingUtils"
 import { SQLBatchTuple } from "@op-engineering/op-sqlite"
 import { useEffect, useState } from "react"
