+++ hooks/datalist/useDatalistModal.tsx
@@ -9,9 +9,9 @@ import EditRouteModal from "@/components/modal/editModal/EditRouteModal"
 import EditStopModal from "@/components/modal/editModal/EditStopModal"
 import EditVehicleTypeModal from "@/components/modal/editModal/EditVehicleTypeModal"
 import { useDialog } from "@/context/DialogContext"
-import { AddableIconType, AddableRoute, AddableStop, AddableVehicleType } from "@/src/types/AddableTravels"
-import { EditableRoute, EditableStop, EditableVehicleType } from "@/src/types/EditableTravels"
-import { Direction, IconType, Stop } from "@/src/types/Travels"
+import { AddableIconType, AddableRoute, AddableStop, AddableVehicleType } from "@/src/types/AddableTypes"
+import { EditableRoute, EditableStop, EditableVehicleType } from "@/src/types/EditableTypes"
+import { Direction, IconType, Stop } from "@/src/types/Types"
 import { useCallback, useState } from "react"
 import useDataOperations from "../data/useDataOperations"
 import useDirections from "../data/useDirections"
